refactor: rename `lib/src` to `lib/shared`
Change-Type: patch
