+++ lib/cli/cli.js
@@ -20,7 +20,7 @@ const _ = require('lodash');
 const fs = require('fs');
 const yargs = require('yargs');
 const utils = require('./utils');
-const EXIT_CODES = require('../src/exit-codes');
+const EXIT_CODES = require('../shared/exit-codes');
 const packageJSON = require('../../package.json');
 
 /**
