+++ lib/cli/etcher.js
@@ -25,7 +25,7 @@ const drivelist = Bluebird.promisifyAll(require('drivelist'));
 const writer = require('./writer');
 const utils = require('./utils');
 const options = require('./cli');
-const EXIT_CODES = require('../src/exit-codes');
+const EXIT_CODES = require('../shared/exit-codes');
 
 isElevated().then((elevated) => {
   if (!elevated) {
