+++ lib/gui/app.js
@@ -24,7 +24,7 @@
 
 var angular = require('angular');
 const electron = require('electron');
-const EXIT_CODES = require('../src/exit-codes');
+const EXIT_CODES = require('../shared/exit-codes');
 
 /* eslint-enable no-var */
