+++ lib/shared/child-writer/index.js
