+++ lib/shared/child-writer/renderer-utils.js
