+++ lib/shared/child-writer/utils.js
