+++ lib/shared/child-writer/writer-proxy.js
