+++ lib/shared/exit-codes.js
